requirements.txt
requests
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djoser
pytils==0.3
django-filter~=21.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
asgiref==3.2.10
Django==3.0.5
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
six~=1.16.0